refactor: refactor middleware
move router definition to middleware
move router logic to controllers
load routers and error handlers via app.use

**RELATES TO**: ON-64
