+++ src/middleware/errorHandler.js
@@ -7,38 +7,41 @@
 /**
  * Module dependencies.
  */
+const express = require('express');
+
+const app = express();
+
 const { OpError } = require('../utils').errors;
 
 /**
  * Handles unsupported endpoints and operational errors
  * @param {function} app
  */
-const errorHandler = (app) => {
-  // Unsupported endpoints
-  app.use(function pageNotFoundHandler(req, res) {
-    const error = {
-      id: 'PAGE_NOT_FOUND',
-      status: 404,
-      message: 'Unsupported Endpoint',
-    };
-    res.status(404).json(error);
-  });
 
-  // Other operational errors
-  // eslint-disable-next-line no-unused-vars
-  app.use(function opErrorHandler(err, req, res, next) {
-    if (err instanceof OpError) {
-      const { error } = err;
-      return res.status(error.statusCode).json(error);
-    }
+// Unsupported endpoints
+app.use(function pageNotFoundHandler(req, res) {
+  const error = {
+    id: 'PAGE_NOT_FOUND',
+    status: 404,
+    message: 'Unsupported Endpoint',
+  };
+  res.status(404).json(error);
+});
 
-    const error = {
-      id: 'UNIDENTIFIED_ERROR',
-      status: 500,
-      message: 'Unidentified server error',
-    };
-    return res.status(500).json(error);
-  });
-};
+// Other operational errors
+// eslint-disable-next-line no-unused-vars
+app.use(function opErrorHandler(err, req, res, next) {
+  if (err instanceof OpError) {
+    const { error } = err;
+    return res.status(error.statusCode).json(error);
+  }
 
-module.exports = errorHandler;
+  const error = {
+    id: 'UNIDENTIFIED_ERROR',
+    status: 500,
+    message: 'Unidentified server error',
+  };
+  return res.status(500).json(error);
+});
+
+module.exports = app;
