+++ src/middleware/routers/beneficiaries.js
@@ -10,7 +10,9 @@
 const express = require('express');
 
 const router = express.Router();
-const { beneficiaries } = require('../controllers');
+const {
+  router: { beneficiaries },
+} = require('../../controllers');
 
 /**
  * Routing for beneficiaries endpoints (/v1/beneficiaries)
